Refactor Tacticus Guild Raid Integration
- Consolidated column definitions and renderers into separate files for better organization and maintainability.
- Moved user summary and token management logic to a dedicated model file.
- Updated the main visualization component to utilize the new structure, improving readability and reducing complexity.
- Removed unused imports and code, streamlining the component.
- Enhanced token and bomb status rendering logic for clarity and efficiency.
